aider: Updated `autostart_app.desktop.j2` template to render settings from a dictionary in the format `Key=Value`.
# Aider chat conversation:

USER: Is there a way the template can be modified to accept a dict where the key is the name of a setting and the value is the value of the setting?

ASSISTANT: To modify the `autostart_app.desktop.j2` template to accept a dictionary where the key is the name of a setting and the value is the value of the setting, we can iterate over the dictionary items and render each key-value pair in the template. Here's the diff for the changes:

```diff
--- vm-setup/roles/gnome_settings/templates/autostart_app.desktop.j2
+++ vm-setup/roles/gnome_settings/templates/autostart_app.desktop.j2
@@ ... @@
 [Desktop Entry]
-Type=Application
-{% if autostart_app_exec is defined and autostart_app_exec | trim != "" %}
-Exec={{ autostart_app_exec }}
-{% endif %}
-{% if autostart_app_path is defined and autostart_app_path | trim != "" %}
-Path={{ autostart_app_path }}
-{% endif %}
-Hidden={{ autostart_app_hidden }}
-{% if autostart_app_comment is defined and autostart_app_comment | trim != "" %}
-Comment={{ autostart_app_comment }}
-{% endif %}
-{% if autostart_app_terminal is defined %}
-Terminal={{ autostart_app_terminal | lower }}
-{% endif %}
-X-GNOME-Autostart-enabled=true
+{% for key, value in autostart_app_settings.items() %}
+{{ key }}={{ value }}
+{% endfor %}
```

This change assumes that you have a variable `autostart_app_settings` which is a dictionary containing the settings for the desktop entry. Each item in the dictionary will be rendered as a line in the format `Key=Value`.

Please ensure that the `autostart_app_settings` dictionary is defined in your Ansible variables and contains all the necessary keys and values for the desktop entry.