aider: Added the UUID for the `extension-list` GNOME Shell extension to the `vars/main.yml` file.
# Aider chat conversation:

USER: Add the uuid for extension-list to the vars/main.yml as a list to be passed to the block just created. add a comment with the name of the extension.

ASSISTANT: To add the UUID for the `extension-list` GNOME Shell extension to the `vars/main.yml` file in the `gnome_settings` role, and to pass it to the block that was created previously, here is the diff:

```diff
--- vm-setup/roles/gnome_settings/vars/main.yml
+++ vm-setup/roles/gnome_settings/vars/main.yml
@@ ... @@
 gnome_packages:
   - gnome-shell-extension-desktop-icons-ng
   - gnome-shell-extensions
   - gnome-startup-applications
   - gnome-tweaks
   - tilix
   - wmctrl
   - gnome-shell-extension-installer
+gnome_shell_extension_uuids:
+  # Extension List - Provides a list of installed GNOME Shell extensions
+  - "extension-list@gnome-shell-extensions.gcampax.github.com"
```

This change adds a new variable `gnome_shell_extension_uuids` as a list containing the UUID for the `extension-list` extension, with a comment describing the extension. This list can then be used in the block that installs and enables GNOME Shell extensions.