aider: Remove `Hidden` key from Tilix task settings.
# Aider chat conversation:

USER: Is the hidden setting required in the tilix task?

ASSISTANT: The `Hidden` setting in a `.desktop` file is used to specify whether the application should be hidden from menus or not. By default, if this key is not provided, the application is not hidden. Since the `Hidden` key is not required for the `.desktop` file to function correctly, especially if you want the application to be visible, it can be omitted unless you specifically want to hide the application.

If you want to remove the `Hidden` setting from the Tilix task in `tilix.yml`, here's the diff for that change:

```diff
--- vm-setup/roles/gnome_settings/tasks/tilix.yml
+++ vm-setup/roles/gnome_settings/tasks/tilix.yml
@@ ... @@
 - name: Include autostart creation tasks
   include_tasks: create_autostart.yml
   vars:
     autostart_app_type: "Application"
     autostart_app_name: "Tilix"
     autostart_app_exec: "tilix"
     autostart_app_settings:
       # These keys are case sensitive. Make sure you're using the correct
       # capitalization for these key.
       Comment: "Tilix terminal emulator"
-      Hidden: false
       X-GNOME-Autostart-enabled: true
       X-GNOME-Autostart-Delay: 0
```

This change will remove the `Hidden` key from the settings that are passed to the `create_autostart.yml` task, which means the generated `.desktop` file for Tilix will not have the `Hidden` key and the application will be visible in menus.